refac(file path utils): move functions to seperate module
